refactor: split lib.rs into 5 modules (vad, geo, vtec, nwws, utils)
Rust module organization:
- utils.rs: sum_as_string, calculate_fast_hash, validate_image_cache_batch, normalize_product_id
- vad.rs: VWP/VAD parser, hodograph math (vec2comp, comp2vec, compute_bunkers, etc.)
- geo.rs: R-tree radar lookup, haversine distance, point-in-polygon
- vtec.rs: VTEC string parser, warning polygon extraction, narrative cleaning
- nwws.rs: XMPP client sidecar, S3 VAD fetch

All functions exported via lib.rs module declarations.
