+++ src_rust/vad.rs
@@ -1,0 +1,717 @@
+use pyo3::prelude::*;
+use pyo3::types::PyDict;
+
+// ── find_vwp_header_offset ──
+
+#[pyfunction]
+pub fn find_vwp_header_offset(data: &[u8]) -> PyResult<Option<usize>> {
+    if data.len() < 2 {
+        return Ok(None);
+    }
+    let search_limit = std::cmp::min(data.len() - 2, 200);
+    for i in 0..=search_limit {
+        if data[i] == 0x00 && data[i + 1] == 0x30 {
+            if i >= 30 {
+                if data[i - 30] == 0x00 && data[i - 30 + 1] == 0x30 {
+                    return Ok(Some(i));
+                }
+            }
+        }
+    }
+    Ok(None)
+}
+
+// ── VadRecord+parse_vwp_tabular_data ──
+
+struct VadRecord {
+    wind_dir: f64,
+    wind_spd: f64,
+    rms_error: f64,
+    divergence: f64,
+    slant_range: f64,
+    elev_angle: f64,
+    altitude: f64,
+}
+
+/// Fast parser for VWP tabular data.
+/// Returns a dictionary of lists (wind_dir, wind_spd, etc.) sorted by altitude.
+/// Processes all pages of multi-page VAD data and merges results.
+#[pyfunction]
+pub fn parse_vwp_tabular_data<'py>(
+    py: Python<'py>,
+    data: &[u8],
+    _offset_tabular: usize,
+) -> PyResult<Option<Bound<'py, PyDict>>> {
+    let marker = b"VAD Algorithm Output";
+    let mut records = Vec::new();
+    let mut search_start = 0;
+
+    // Find all occurrences of marker (handles multi-page VAD files)
+    while let Some(marker_pos) = data[search_start..]
+        .windows(marker.len())
+        .position(|w| w == marker)
+    {
+        let actual_pos = search_start + marker_pos;
+
+        // Find the 0x00 0x50 line start marker before this data marker
+        let mut line_start = 0;
+        for i in (0..actual_pos).rev() {
+            if i + 2 <= data.len() && data[i] == 0x00 && data[i + 1] == 0x50 {
+                line_start = i;
+                break;
+            }
+        }
+
+        if line_start == 0 {
+            search_start = actual_pos + marker.len();
+            if search_start >= data.len() {
+                break;
+            }
+            continue;
+        }
+
+        // Parse this page starting from line_start
+        let mut current_pos = line_start;
+        let mut line_count = 0;
+
+        while current_pos + 82 <= data.len() {
+            let len = i16::from_be_bytes([data[current_pos], data[current_pos + 1]]);
+
+            if len != 80 {
+                if len == -1 || len == 0 {
+                    break;
+                }
+                current_pos += 2;
+                continue;
+            }
+
+            // Skip first 3 lines of headers
+            if line_count >= 3 {
+                let line_bytes = &data[current_pos + 2..current_pos + 82];
+                // Use lossy UTF-8 conversion for just this line
+                let line = String::from_utf8_lossy(line_bytes);
+                let parts: Vec<&str> = line.split_whitespace().collect();
+
+                if parts.len() >= 10 {
+                    if let (Some(d), Some(s), Some(r), Some(v), Some(sl), Some(e)) = (
+                        parts[4].parse::<f64>().ok(),
+                        parts[5].parse::<f64>().ok(),
+                        parts[6].parse::<f64>().ok(),
+                        if parts[7] == "NA" {
+                            Some(f64::NAN)
+                        } else {
+                            parts[7].parse::<f64>().ok()
+                        },
+                        parts[8].parse::<f64>().ok(),
+                        parts[9].parse::<f64>().ok(),
+                    ) {
+                        let slant_km: f64 = sl * (6067.1 / 3281.0);
+                        let r_e: f64 = (4.0 / 3.0) * 6371.0;
+                        let elev_rad: f64 = e.to_radians();
+                        let alt = (r_e.powi(2)
+                            + slant_km.powi(2)
+                            + 2.0 * r_e * slant_km * elev_rad.sin())
+                        .sqrt()
+                            - r_e;
+
+                        records.push(VadRecord {
+                            wind_dir: d,
+                            wind_spd: s,
+                            rms_error: r,
+                            divergence: v,
+                            slant_range: slant_km,
+                            elev_angle: e,
+                            altitude: alt,
+                        });
+                    }
+                }
+            }
+
+            current_pos += 82;
+            line_count += 1;
+        }
+
+        search_start = actual_pos + marker.len();
+        if search_start >= data.len() {
+            break;
+        }
+    }
+
+    if records.is_empty() {
+        return Ok(None);
+    }
+
+    // Sort by altitude (all pages merged and sorted together)
+    records.sort_by(|a, b| {
+        a.altitude
+            .partial_cmp(&b.altitude)
+            .unwrap_or(std::cmp::Ordering::Equal)
+    });
+
+    let dict = PyDict::new(py);
+    let mut wind_dir = Vec::with_capacity(records.len());
+    let mut wind_spd = Vec::with_capacity(records.len());
+    let mut rms_error = Vec::with_capacity(records.len());
+    let mut divergence = Vec::with_capacity(records.len());
+    let mut slant_range = Vec::with_capacity(records.len());
+    let mut elev_angle = Vec::with_capacity(records.len());
+    let mut altitude = Vec::with_capacity(records.len());
+
+    for r in records {
+        wind_dir.push(r.wind_dir);
+        wind_spd.push(r.wind_spd);
+        rms_error.push(r.rms_error);
+        divergence.push(r.divergence);
+        slant_range.push(r.slant_range);
+        elev_angle.push(r.elev_angle);
+        altitude.push(r.altitude);
+    }
+
+    dict.set_item("wind_dir", wind_dir)?;
+    dict.set_item("wind_spd", wind_spd)?;
+    dict.set_item("rms_error", rms_error)?;
+    dict.set_item("divergence", divergence)?;
+    dict.set_item("slant_range", slant_range)?;
+    dict.set_item("elev_angle", elev_angle)?;
+    dict.set_item("altitude", altitude)?;
+    Ok(Some(dict))
+}
+
+// ── vec2comp ──
+
+#[pyfunction]
+pub fn vec2comp(wdir: f64, wspd: f64) -> PyResult<(f64, f64)> {
+    let wdir_rad = wdir.to_radians();
+    let u = -wspd * wdir_rad.sin();
+    let v = -wspd * wdir_rad.cos();
+    Ok((u, v))
+}
+
+// ── comp2vec ──
+
+#[pyfunction]
+pub fn comp2vec(u: f64, v: f64) -> PyResult<(f64, f64)> {
+    let spd = (u * u + v * v).sqrt();
+    let dir = if spd > 0.0 {
+        let angle_rad = (-v).atan2(-u);
+        let angle_deg = angle_rad.to_degrees();
+        let dir_deg = (90.0 - angle_deg) % 360.0;
+        if dir_deg < 0.0 {
+            dir_deg + 360.0
+        } else {
+            dir_deg
+        }
+    } else {
+        0.0
+    };
+    Ok((dir, spd))
+}
+
+// ── _clip_profile_internal ──
+
+pub(crate) fn _clip_profile_internal(
+    wind_dir: &[f64],
+    wind_spd: &[f64],
+    altitude: &[f64],
+    max_hght: f64,
+) -> (Vec<f64>, Vec<f64>, Vec<f64>) {
+    let mut clipped_dir = Vec::new();
+    let mut clipped_spd = Vec::new();
+    let mut clipped_alt = Vec::new();
+    for (i, &alt) in altitude.iter().enumerate() {
+        if alt <= max_hght {
+            clipped_dir.push(wind_dir[i]);
+            clipped_spd.push(wind_spd[i]);
+            clipped_alt.push(alt);
+        }
+    }
+    (clipped_dir, clipped_spd, clipped_alt)
+}
+
+// ── _interp_linear ──
+
+pub(crate) fn _interp_linear(x: &[f64], y: &[f64], xi: f64) -> f64 {
+    if x.is_empty() || xi.is_nan() {
+        return f64::NAN;
+    }
+    // Match numpy.interp(left=nan, right=nan)
+    if xi < x[0] || xi > x[x.len() - 1] {
+        return f64::NAN;
+    }
+    if xi == x[0] {
+        return y[0];
+    }
+    if xi == x[x.len() - 1] {
+        return y[y.len() - 1];
+    }
+
+    for i in 0..x.len() - 1 {
+        if x[i] <= xi && xi <= x[i + 1] {
+            let t = (xi - x[i]) / (x[i + 1] - x[i]);
+            return y[i] + t * (y[i + 1] - y[i]);
+        }
+    }
+    f64::NAN
+}
+
+// ── compute_bunkers ──
+
+#[pyfunction]
+pub fn compute_bunkers(
+    wind_dir: Vec<f64>,
+    wind_spd: Vec<f64>,
+    altitude: Vec<f64>,
+) -> PyResult<((f64, f64), (f64, f64), (f64, f64))> {
+    if wind_dir.is_empty() {
+        return Ok((
+            (f64::NAN, f64::NAN),
+            (f64::NAN, f64::NAN),
+            (f64::NAN, f64::NAN),
+        ));
+    }
+
+    let hght = 6.0; // Height in same units as altitude (km)
+    let d = 7.5 * 1.94;
+
+    // Convert all wind vectors to u/v components
+    let mut u_all = Vec::new();
+    let mut v_all = Vec::new();
+    for i in 0..wind_dir.len() {
+        let (u, v) = vec2comp(wind_dir[i], wind_spd[i])?;
+        u_all.push(u);
+        v_all.push(v);
+    }
+
+    // Interpolate u/v to the target height (6 in same units as altitude)
+    let u_hght = _interp_linear(&altitude, &u_all, hght);
+    let v_hght = _interp_linear(&altitude, &v_all, hght);
+
+    // Clip profile: find where altitude[i] <= hght < altitude[i+1]
+    // Then include all points up to i and append the interpolated value
+    let mut clip_idx: Option<usize> = None;
+    for i in 0..(altitude.len() - 1) {
+        if altitude[i] <= hght && hght < altitude[i + 1] {
+            clip_idx = Some(i);
+            break;
+        }
+    }
+
+    let mut u_clip = Vec::new();
+    let mut v_clip = Vec::new();
+
+    if let Some(idx) = clip_idx {
+        // Normal case: hght is between two points
+        for i in 0..=idx {
+            u_clip.push(u_all[i]);
+            v_clip.push(v_all[i]);
+        }
+    } else if !altitude.is_empty() && altitude[altitude.len() - 1] <= hght {
+        // Edge case: hght is at or beyond the last point
+        // Include all points
+        for i in 0..altitude.len() {
+            if altitude[i] <= hght {
+                u_clip.push(u_all[i]);
+                v_clip.push(v_all[i]);
+            }
+        }
+    } else {
+        // hght is below all points or other edge case - return NaN
+        return Ok((
+            (f64::NAN, f64::NAN),
+            (f64::NAN, f64::NAN),
+            (f64::NAN, f64::NAN),
+        ));
+    }
+
+    // Append interpolated value at target height
+    u_clip.push(u_hght);
+    v_clip.push(v_hght);
+
+    if u_clip.len() < 2 {
+        return Ok((
+            (f64::NAN, f64::NAN),
+            (f64::NAN, f64::NAN),
+            (f64::NAN, f64::NAN),
+        ));
+    }
+
+    // Mean wind 0-6km: average of u/v components
+    let mnu6 = u_clip.iter().sum::<f64>() / u_clip.len() as f64;
+    let mnv6 = v_clip.iter().sum::<f64>() / v_clip.len() as f64;
+
+    // Shear: difference between top (interpolated at hght) and bottom (surface)
+    let shru = u_hght - u_all[0];
+    let shrv = v_hght - v_all[0];
+
+    // Bunkers displacement: perpendicular to shear
+    let shear_mag = (shru * shru + shrv * shrv).sqrt();
+    let tmp = if shear_mag > 0.01 { d / shear_mag } else { 0.0 };
+
+    // Right and left storm motions
+    let rstu = mnu6 + (tmp * shrv);
+    let rstv = mnv6 - (tmp * shru);
+    let lstu = mnu6 - (tmp * shrv);
+    let lstv = mnv6 + (tmp * shru);
+
+    // Convert back to dir/spd
+    let (right_dir, right_spd) = comp2vec(rstu, rstv)?;
+    let (left_dir, left_spd) = comp2vec(lstu, lstv)?;
+    let (mean_dir, mean_spd) = comp2vec(mnu6, mnv6)?;
+
+    // Return in Python order: (right, left, mean)
+    Ok((
+        (right_dir, right_spd),
+        (left_dir, left_spd),
+        (mean_dir, mean_spd),
+    ))
+}
+
+// ── compute_srh ──
+
+#[pyfunction]
+pub fn compute_srh(
+    wind_dir: Vec<f64>,
+    wind_spd: Vec<f64>,
+    altitude: Vec<f64>,
+    storm_dir: f64,
+    storm_spd: f64,
+    hght_km: f64,
+) -> PyResult<f64> {
+    if wind_dir.is_empty() {
+        return Ok(f64::NAN);
+    }
+
+    let hght = hght_km; // Altitude is already in km
+
+    // Convert all wind vectors to u/v components
+    let mut u_all = Vec::new();
+    let mut v_all = Vec::new();
+    for i in 0..wind_dir.len() {
+        let (u, v) = vec2comp(wind_dir[i], wind_spd[i])?;
+        u_all.push(u);
+        v_all.push(v);
+    }
+
+    // Convert storm motion to u/v
+    let (storm_u, storm_v) = vec2comp(storm_dir, storm_spd)?;
+
+    // Compute storm-relative wind for entire profile
+    let mut sru = Vec::new();
+    let mut srv = Vec::new();
+    for i in 0..u_all.len() {
+        sru.push((u_all[i] - storm_u) / 1.94);
+        srv.push((v_all[i] - storm_v) / 1.94);
+    }
+
+    // Interpolate to target height
+    let sru_hght = _interp_linear(&altitude, &sru, hght);
+    let srv_hght = _interp_linear(&altitude, &srv, hght);
+
+    // Clip and append interpolated values (matching Python's _clip_profile)
+    let mut sru_clip = Vec::new();
+    let mut srv_clip = Vec::new();
+    for i in 0..altitude.len() {
+        if altitude[i] <= hght {
+            sru_clip.push(sru[i]);
+            srv_clip.push(srv[i]);
+        }
+    }
+    sru_clip.push(sru_hght);
+    srv_clip.push(srv_hght);
+
+    if sru_clip.len() < 2 {
+        return Ok(f64::NAN);
+    }
+
+    // Compute cross products between consecutive layers: u[i+1]*v[i] - u[i]*v[i+1]
+    let mut srh = 0.0;
+    for i in 0..(sru_clip.len() - 1) {
+        let cross = (sru_clip[i + 1] * srv_clip[i]) - (sru_clip[i] * srv_clip[i + 1]);
+        srh += cross;
+    }
+
+    Ok(srh)
+}
+
+// ── compute_critical_angle ──
+
+#[pyfunction]
+pub fn compute_critical_angle(
+    wind_dir: Vec<f64>,
+    wind_spd: Vec<f64>,
+    altitude: Vec<f64>,
+    storm_dir: f64,
+    storm_spd: f64,
+) -> PyResult<f64> {
+    if wind_dir.is_empty() {
+        return Ok(f64::NAN);
+    }
+
+    let (clipped_dir, clipped_spd, _) =
+        _clip_profile_internal(&wind_dir, &wind_spd, &altitude, 6000.0);
+    if clipped_dir.is_empty() {
+        return Ok(f64::NAN);
+    }
+
+    let storm_rad = storm_dir.to_radians();
+    let (storm_u, storm_v) = (-storm_spd * storm_rad.sin(), -storm_spd * storm_rad.cos());
+
+    let mut min_angle: f64 = 360.0;
+    for i in 0..clipped_dir.len() {
+        let wd = clipped_dir[i].to_radians();
+        let u = -clipped_spd[i] * wd.sin();
+        let v = -clipped_spd[i] * wd.cos();
+
+        let rel_u = u - storm_u;
+        let rel_v = v - storm_v;
+        let rel_dir = ((-rel_v).atan2(-rel_u).to_degrees() + 90.0) % 360.0;
+        let diff = (rel_dir - storm_dir).abs();
+        if diff > 180.0 {
+            min_angle = min_angle.min(360.0 - diff);
+        } else {
+            min_angle = min_angle.min(diff);
+        }
+    }
+
+    Ok(min_angle)
+}
+
+// ── compute_dtm ──
+
+#[pyfunction]
+pub fn compute_dtm(
+    wind_dir: Vec<f64>,
+    wind_spd: Vec<f64>,
+    altitude: Vec<f64>,
+) -> PyResult<(f64, f64)> {
+    if wind_dir.is_empty() {
+        return Ok((f64::NAN, f64::NAN));
+    }
+
+    // Convert all wind to u/v
+    let mut u_all = Vec::with_capacity(wind_dir.len());
+    let mut v_all = Vec::with_capacity(wind_dir.len());
+    for i in 0..wind_dir.len() {
+        let (u, v) = vec2comp(wind_dir[i], wind_spd[i])?;
+        u_all.push(u);
+        v_all.push(v);
+    }
+
+    // Mean u/v over 0–0.5 km using 20 linearly-spaced interpolation points
+    let n = 20usize;
+    let mut sum_u = 0.0f64;
+    let mut sum_v = 0.0f64;
+    let mut count = 0usize;
+    for k in 0..n {
+        let h = 0.5 * k as f64 / (n - 1) as f64;
+        let ui = _interp_linear(&altitude, &u_all, h);
+        let vi = _interp_linear(&altitude, &v_all, h);
+        if ui.is_finite() && vi.is_finite() {
+            sum_u += ui;
+            sum_v += vi;
+            count += 1;
+        }
+    }
+    if count == 0 {
+        return Ok((f64::NAN, f64::NAN));
+    }
+    let mn_u_500 = sum_u / count as f64;
+    let mn_v_500 = sum_v / count as f64;
+
+    // Bunkers right-mover
+    let ((brm_dir, brm_spd), _, _) = compute_bunkers(wind_dir, wind_spd, altitude)?;
+    if brm_dir.is_nan() {
+        return Ok((f64::NAN, f64::NAN));
+    }
+    let (brm_u, brm_v) = vec2comp(brm_dir, brm_spd)?;
+
+    // DTM = 0.7 * BRM + 0.3 * mean_0500m
+    let dtm_u = 0.7 * brm_u + 0.3 * mn_u_500;
+    let dtm_v = 0.7 * brm_v + 0.3 * mn_v_500;
+
+    comp2vec(dtm_u, dtm_v)
+}
+
+// ── compute_crit_angl ──
+
+#[pyfunction]
+pub fn compute_crit_angl(
+    wind_dir: Vec<f64>,
+    wind_spd: Vec<f64>,
+    altitude: Vec<f64>,
+    storm_dir: f64,
+    storm_spd: f64,
+) -> PyResult<f64> {
+    if wind_dir.is_empty() {
+        return Ok(f64::NAN);
+    }
+
+    let (storm_u, storm_v) = vec2comp(storm_dir, storm_spd)?;
+
+    // Surface u/v
+    let (u0, v0) = vec2comp(wind_dir[0], wind_spd[0])?;
+
+    // Convert all to u/v for interpolation
+    let mut u_all = Vec::with_capacity(wind_dir.len());
+    let mut v_all = Vec::with_capacity(wind_dir.len());
+    for i in 0..wind_dir.len() {
+        let (u, v) = vec2comp(wind_dir[i], wind_spd[i])?;
+        u_all.push(u);
+        v_all.push(v);
+    }
+
+    // Interpolate to 0.5 km
+    let u_05 = _interp_linear(&altitude, &u_all, 0.5);
+    let v_05 = _interp_linear(&altitude, &v_all, 0.5);
+    if u_05.is_nan() || v_05.is_nan() {
+        return Ok(f64::NAN);
+    }
+
+    // base = storm motion relative to surface wind
+    let base_u = storm_u - u0;
+    let base_v = storm_v - v0;
+
+    // ang = 0.5km wind relative to surface wind (0–0.5km shear)
+    let ang_u = u_05 - u0;
+    let ang_v = v_05 - v0;
+
+    let len_base = (base_u * base_u + base_v * base_v).sqrt();
+    let len_ang = (ang_u * ang_u + ang_v * ang_v).sqrt();
+
+    if len_base < 1e-9 || len_ang < 1e-9 {
+        return Ok(f64::NAN);
+    }
+
+    let cos_theta = (base_u * ang_u + base_v * ang_v) / (len_base * len_ang);
+    // Clamp to [-1, 1] to guard against float rounding past the domain of acos
+    let cos_theta = cos_theta.clamp(-1.0, 1.0);
+    Ok(cos_theta.acos().to_degrees())
+}
+
+// ── linspace ──
+
+pub(crate) fn linspace(start: f64, end: f64, n: usize) -> Vec<f64> {
+    if n == 0 {
+        return Vec::new();
+    }
+    if n == 1 {
+        return vec![start];
+    }
+    let step = (end - start) / (n - 1) as f64;
+    (0..n).map(|i| start + i as f64 * step).collect()
+}
+
+// ── compute_shear_mag ──
+
+#[pyfunction]
+pub fn compute_shear_mag(
+    wind_dir: Vec<f64>,
+    wind_spd: Vec<f64>,
+    altitude: Vec<f64>,
+    hght: f64,
+) -> PyResult<f64> {
+    if wind_dir.is_empty() || altitude.is_empty() {
+        return Ok(f64::NAN);
+    }
+
+    let mut u_all = Vec::with_capacity(wind_dir.len());
+    let mut v_all = Vec::with_capacity(wind_dir.len());
+    for i in 0..wind_dir.len() {
+        let (u, v) = vec2comp(wind_dir[i], wind_spd[i])?;
+        u_all.push(u);
+        v_all.push(v);
+    }
+
+    let u_hght = _interp_linear(&altitude, &u_all, hght);
+    let v_hght = _interp_linear(&altitude, &v_all, hght);
+
+    if u_hght.is_nan() || v_hght.is_nan() {
+        return Ok(f64::NAN);
+    }
+
+    let du = u_hght - u_all[0];
+    let dv = v_hght - v_all[0];
+    Ok((du * du + dv * dv).sqrt())
+}
+
+// ── compute_sr_flow ──
+
+#[pyfunction]
+pub fn compute_sr_flow(
+    wind_dir: Vec<f64>,
+    wind_spd: Vec<f64>,
+    altitude: Vec<f64>,
+    storm_dir: f64,
+    storm_spd: f64,
+    hght_bot: f64,
+    hght_top: f64,
+) -> PyResult<f64> {
+    if wind_dir.is_empty() || altitude.is_empty() {
+        return Ok(f64::NAN);
+    }
+
+    let mut u_all = Vec::with_capacity(wind_dir.len());
+    let mut v_all = Vec::with_capacity(wind_dir.len());
+    for i in 0..wind_dir.len() {
+        let (u, v) = vec2comp(wind_dir[i], wind_spd[i])?;
+        u_all.push(u);
+        v_all.push(v);
+    }
+
+    let (storm_u, storm_v) = vec2comp(storm_dir, storm_spd)?;
+
+    let n = 50;
+    let layer_alts = linspace(hght_bot, hght_top, n);
+
+    let mut sum_mag = 0.0;
+    let mut count = 0;
+
+    for alt in layer_alts {
+        let u_interp = _interp_linear(&altitude, &u_all, alt);
+        let v_interp = _interp_linear(&altitude, &v_all, alt);
+
+        if u_interp.is_finite() && v_interp.is_finite() {
+            let sr_u = u_interp - storm_u;
+            let sr_v = v_interp - storm_v;
+            sum_mag += (sr_u * sr_u + sr_v * sr_v).sqrt();
+            count += 1;
+        }
+    }
+
+    if count == 0 {
+        Ok(f64::NAN)
+    } else {
+        Ok(sum_mag / count as f64)
+    }
+}
+
+// ── clip_profile ──
+
+#[pyfunction]
+pub fn clip_profile(
+    prof: Vec<f64>,
+    alt: Vec<f64>,
+    clip_alt: f64,
+    intrp_prof: f64,
+) -> PyResult<Vec<f64>> {
+    if alt.len() < 2 {
+        return Ok(vec![f64::NAN; prof.len()]);
+    }
+
+    let mut idx_clip = None;
+    for i in 0..(alt.len() - 1) {
+        if alt[i] <= clip_alt && alt[i + 1] > clip_alt {
+            idx_clip = Some(i);
+            break;
+        }
+    }
+
+    match idx_clip {
+        Some(idx) => {
+            let mut result = prof[0..=idx].to_vec();
+            result.push(intrp_prof);
+            Ok(result)
+        }
+        None => Ok(vec![f64::NAN; prof.len()]),
+    }
+}
